fix: proper type definitions
